rollingpolicy: Don't crash on redefinition before initialization [BAT-10497]
The rolling policy initialization function gets called either when the first
appender uses it or when a redefinition occurs (with the intent of updating its
values). In the latter case it would get called with a null rollingfile_udata_t
pointer and then crash.

The solution is simply to ignore the call to the init function if the udata
pointer is null. It'll get called again on the first use, at which time the
udata pointer will be valid.

An easy way to reproduce was to have a config section like this:

  <!-- Definition -->
  <rollingpolicy name="mypolicy" type="sizewin" maxsize="1024" maxnum="10" />
  <!-- Re-definition, calls log4c_rollingpolicy_init() and crashes -->
  <rollingpolicy name="mypolicy" type="sizewin" maxsize="2048" maxnum="5" />

  <!-- First use -->
  <appender name="myappender" type="rollingfile" logdir="." layout="dated" rollingpolicy="mypolicy" />

This, on the other hand, would not crash:

  <!-- Definition -->
  <rollingpolicy name="mypolicy" type="sizewin" maxsize="1024" maxnum="10" />

  <!-- First use, calls log4c_rollingpolicy_init() -->
  <appender name="myappender" type="rollingfile" logdir="." layout="dated" rollingpolicy="mypolicy" />

  <!-- Re-definition, calls log4c_rollingpolicy_init() again -->
  <rollingpolicy name="mypolicy" type="sizewin" maxsize="2048" maxnum="5" />

The issue was particularly annoying when copying a global log4crc (e.g. the one
in ${LOG4C_RCPATH}/log4crc) to the home directory (${HOME}/.log4crc), hoping to
override the former. Any executable using the log service would then crash.