Allow config files of an older version to be read
The previous version comparison was unnecessarily strict and even a micro
version bump (e.g. from 1.2.1 to 1.2.2) would cause log4c to more or less
silently ignore a log4crc file that contained an older version tag.

With this change we do semantic version comparison and only reject newer config
files. If the configuration file format ever changes to be incompatible we can
change the logic.